fix(d3d12): widen index and container arithmetic to 64 bits
StartIndexLocation is app-controlled and unbounded, but both indexed-draw
encoders multiplied it by the index size in UINT before assigning to a UINT64
byte offset. With R32 indices a start location of 0x40000000 wraps the product
back to 0, so the draw silently reads the wrong indices from inside the buffer
and passes every downstream bounds check. d3d12_draw_state_audit.cpp already
did this computation in uint64_t; the two encode paths were the outliers.

FindDxbcShaderBlob checked only that the caller's buffer was at least 32 bytes,
never that the container claimed to be. A container_size below 32 wrapped
container_size - 32 to roughly 4G, defeating the blob-table budget check, and
the table walk then read far past the caller's buffer. The bytecode comes
straight from D3D12_SHADER_BYTECODE, so the input is entirely app-controlled.

The remaining two sites are capacity hints whose operands are only bounded by
UINT16 descriptor fields; widen them rather than rely on that.

+++ src/d3d12/d3d12_replay_binding_types.hpp
@@ -94,7 +94,8 @@ struct SubmittedDescriptorRecordStore {
       const SubmittedDescriptorRecordKey key = {
           record.mirror, record.heap_index, record.slot_version};
       if (heap_record_slots.empty() ||
-          (heap_record_count + 1) * 2 > heap_record_slots.size())
+          (static_cast<size_t>(heap_record_count) + 1) * 2 >
+              heap_record_slots.size())
         reserveHeapRecords(std::max<size_t>(16, heap_record_count + 1));
       auto &slot = findHeapRecordSlot(key);
       if (slot.index_plus_one) {
